--- baseapp/utils_test.go
@@ -1,279 +1,0 @@
-package baseapp_test
-
-import (
-	"bytes"
-	"context"
-	"encoding/binary"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"net/url"
-	"reflect"
-	"strconv"
-	"testing"
-	"unsafe"
-
-	cmtproto "github.com/cometbft/cometbft/proto/tendermint/types"
-	dbm "github.com/cosmos/cosmos-db"
-	"github.com/stretchr/testify/require"
-
-	runtimev1alpha1 "cosmossdk.io/api/cosmos/app/runtime/v1alpha1"
-	appv1alpha1 "cosmossdk.io/api/cosmos/app/v1alpha1"
-	"cosmossdk.io/core/address"
-	"cosmossdk.io/depinject"
-	"cosmossdk.io/depinject/appconfig"
-	errorsmod "cosmossdk.io/errors"
-	storetypes "cosmossdk.io/store/types"
-
-	"github.com/cosmos/cosmos-sdk/baseapp"
-	baseapptestutil "github.com/cosmos/cosmos-sdk/baseapp/testutil"
-	"github.com/cosmos/cosmos-sdk/client"
-	addresscodec "github.com/cosmos/cosmos-sdk/codec/address"
-	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/cosmos/cosmos-sdk/types/mempool"
-	signingtypes "github.com/cosmos/cosmos-sdk/types/tx/signing"
-)
-
-var ParamStoreKey = []byte("paramstore")
-
-func makeMinimalConfig() depinject.Config {
-	var (
-		mempoolOpt            = baseapp.SetMempool(mempool.NewSenderNonceMempool())
-		addressCodec          = func() address.Codec { return addresscodec.NewBech32Codec("cosmos") }
-		validatorAddressCodec = func() address.ValidatorAddressCodec { return addresscodec.NewBech32Codec("cosmosvaloper") }
-		consensusAddressCodec = func() address.ConsensusAddressCodec { return addresscodec.NewBech32Codec("cosmosvalcons") }
-	)
-
-	return depinject.Configs(
-		depinject.Supply(mempoolOpt, addressCodec, validatorAddressCodec, consensusAddressCodec),
-		appconfig.Compose(&appv1alpha1.Config{
-			Modules: []*appv1alpha1.ModuleConfig{
-				{
-					Name: "runtime",
-					Config: appconfig.WrapAny(&runtimev1alpha1.Module{
-						AppName: "BaseAppApp",
-					}),
-				},
-			},
-		}))
-}
-
-type MsgKeyValueImpl struct{}
-
-func (m MsgKeyValueImpl) Set(ctx context.Context, msg *baseapptestutil.MsgKeyValue) (*baseapptestutil.MsgCreateKeyValueResponse, error) {
-	sdkCtx := sdk.UnwrapSDKContext(ctx)
-	sdkCtx.KVStore(capKey2).Set(msg.Key, msg.Value)
-	return &baseapptestutil.MsgCreateKeyValueResponse{}, nil
-}
-
-type CounterServerImplGasMeterOnly struct {
-	gas uint64
-}
-
-func (m CounterServerImplGasMeterOnly) IncrementCounter(ctx context.Context, msg *baseapptestutil.MsgCounter) (*baseapptestutil.MsgCreateCounterResponse, error) {
-	sdkCtx := sdk.UnwrapSDKContext(ctx)
-	gas := m.gas
-
-	// if no gas is provided, use the counter as gas. This is useful for testing
-	if gas == 0 {
-		gas = uint64(msg.Counter)
-	}
-
-	sdkCtx.GasMeter().ConsumeGas(gas, "test")
-	return &baseapptestutil.MsgCreateCounterResponse{}, nil
-}
-
-type NoopCounterServerImpl struct{}
-
-func (m NoopCounterServerImpl) IncrementCounter(
-	_ context.Context,
-	_ *baseapptestutil.MsgCounter,
-) (*baseapptestutil.MsgCreateCounterResponse, error) {
-	return &baseapptestutil.MsgCreateCounterResponse{}, nil
-}
-
-type CounterServerImpl struct {
-	t          *testing.T
-	capKey     storetypes.StoreKey
-	deliverKey []byte
-}
-
-func (m CounterServerImpl) IncrementCounter(ctx context.Context, msg *baseapptestutil.MsgCounter) (*baseapptestutil.MsgCreateCounterResponse, error) {
-	return incrementCounter(ctx, m.t, m.capKey, m.deliverKey, msg)
-}
-
-type Counter2ServerImpl struct {
-	t          *testing.T
-	capKey     storetypes.StoreKey
-	deliverKey []byte
-}
-
-func (m Counter2ServerImpl) IncrementCounter(ctx context.Context, msg *baseapptestutil.MsgCounter2) (*baseapptestutil.MsgCreateCounterResponse, error) {
-	return incrementCounter(ctx, m.t, m.capKey, m.deliverKey, msg)
-}
-
-func incrementCounter(ctx context.Context,
-	t *testing.T,
-	capKey storetypes.StoreKey,
-	deliverKey []byte,
-	msg sdk.Msg,
-) (*baseapptestutil.MsgCreateCounterResponse, error) {
-	t.Helper()
-	sdkCtx := sdk.UnwrapSDKContext(ctx)
-	store := sdkCtx.KVStore(capKey)
-
-	sdkCtx.GasMeter().ConsumeGas(5, "test")
-
-	var msgCount int64
-
-	switch m := msg.(type) {
-	case *baseapptestutil.MsgCounter:
-		if m.FailOnHandler {
-			return nil, errorsmod.Wrap(sdkerrors.ErrInvalidRequest, "message handler failure")
-		}
-		msgCount = m.Counter
-	case *baseapptestutil.MsgCounter2:
-		if m.FailOnHandler {
-			return nil, errorsmod.Wrap(sdkerrors.ErrInvalidRequest, "message handler failure")
-		}
-		msgCount = m.Counter
-	}
-
-	sdkCtx.EventManager().EmitEvents(
-		counterEvent(sdk.EventTypeMessage, msgCount),
-	)
-
-	_, err := incrementingCounter(t, store, deliverKey, msgCount)
-	if err != nil {
-		return nil, err
-	}
-
-	return &baseapptestutil.MsgCreateCounterResponse{}, nil
-}
-
-func counterEvent(evType string, msgCount int64) sdk.Events {
-	return sdk.Events{
-		sdk.NewEvent(
-			evType,
-			sdk.NewAttribute("update_counter", fmt.Sprintf("%d", msgCount)),
-		),
-	}
-}
-
-func incrementingCounter(t *testing.T, store storetypes.KVStore, counterKey []byte, counter int64) (*sdk.Result, error) {
-	t.Helper()
-	storedCounter := getIntFromStore(t, store, counterKey)
-	require.Equal(t, storedCounter, counter)
-	setIntOnStore(store, counterKey, counter+1)
-	return &sdk.Result{}, nil
-}
-
-func setIntOnStore(store storetypes.KVStore, key []byte, i int64) {
-	bz := make([]byte, 8)
-	n := binary.PutVarint(bz, i)
-	store.Set(key, bz[:n])
-}
-
-type paramStore struct {
-	db *dbm.MemDB
-}
-
-var _ baseapp.ParamStore = (*paramStore)(nil)
-
-func (ps paramStore) Set(_ context.Context, value cmtproto.ConsensusParams) error {
-	bz, err := json.Marshal(value)
-	if err != nil {
-		return err
-	}
-
-	return ps.db.Set(ParamStoreKey, bz)
-}
-
-func (ps paramStore) Has(_ context.Context) (bool, error) {
-	return ps.db.Has(ParamStoreKey)
-}
-
-func (ps paramStore) Get(_ context.Context) (cmtproto.ConsensusParams, error) {
-	bz, err := ps.db.Get(ParamStoreKey)
-	if err != nil {
-		return cmtproto.ConsensusParams{}, err
-	}
-
-	if len(bz) == 0 {
-		return cmtproto.ConsensusParams{}, errors.New("params not found")
-	}
-
-	var params cmtproto.ConsensusParams
-	if err := json.Unmarshal(bz, &params); err != nil {
-		return cmtproto.ConsensusParams{}, err
-	}
-
-	return params, nil
-}
-
-func setTxSignature(t *testing.T, builder client.TxBuilder, nonce uint64) {
-	t.Helper()
-	privKey := secp256k1.GenPrivKeyFromSecret([]byte("test"))
-	pubKey := privKey.PubKey()
-	err := builder.SetSignatures(
-		signingtypes.SignatureV2{
-			PubKey:   pubKey,
-			Sequence: nonce,
-			Data:     &signingtypes.SingleSignatureData{},
-		},
-	)
-	require.NoError(t, err)
-}
-
-func testLoadVersionHelper(t *testing.T, app *baseapp.BaseApp, expectedHeight int64, expectedID storetypes.CommitID) {
-	t.Helper()
-	lastHeight := app.LastBlockHeight()
-	lastID := app.LastCommitID()
-	require.Equal(t, expectedHeight, lastHeight)
-	require.Equal(t, expectedID, lastID)
-}
-
-func getCheckStateCtx(app *baseapp.BaseApp) sdk.Context {
-	v := reflect.ValueOf(app).Elem()
-	f := v.FieldByName("checkState")
-	rf := reflect.NewAt(f.Type(), unsafe.Pointer(f.UnsafeAddr())).Elem()
-	return rf.MethodByName("Context").Call(nil)[0].Interface().(sdk.Context)
-}
-
-func getFinalizeBlockStateCtx(app *baseapp.BaseApp) sdk.Context {
-	v := reflect.ValueOf(app).Elem()
-	f := v.FieldByName("finalizeBlockState")
-	rf := reflect.NewAt(f.Type(), unsafe.Pointer(f.UnsafeAddr())).Elem()
-	return rf.MethodByName("Context").Call(nil)[0].Interface().(sdk.Context)
-}
-
-func parseTxMemo(t *testing.T, tx sdk.Tx) (counter int64, failOnAnte bool) {
-	t.Helper()
-	txWithMemo, ok := tx.(sdk.TxWithMemo)
-	require.True(t, ok)
-
-	memo := txWithMemo.GetMemo()
-	vals, err := url.ParseQuery(memo)
-	require.NoError(t, err)
-
-	counter, err = strconv.ParseInt(vals.Get("counter"), 10, 64)
-	require.NoError(t, err)
-
-	failOnAnte = vals.Get("failOnAnte") == "true"
-	return counter, failOnAnte
-}
-
-func getIntFromStore(t *testing.T, store storetypes.KVStore, key []byte) int64 {
-	t.Helper()
-	bz := store.Get(key)
-	if len(bz) == 0 {
-		return 0
-	}
-
-	i, err := binary.ReadVarint(bytes.NewBuffer(bz))
-	require.NoError(t, err)
-
-	return i
-}
