+++ tests/e2e/baseapp/streaming_test.go
@@ -84,7 +84,7 @@ func TestABCI_MultiListener_StateChanges(t *testing.T) {
 
 			sKey := []byte(fmt.Sprintf("distKey%d", i))
 			sVal := []byte(fmt.Sprintf("distVal%d", i))
-			store := getFinalizeBlockStateCtx(suite.baseApp).KVStore(distKey1)
+			store := baseapptestutil.GetFinalizeBlockStateCtx(suite.baseApp).KVStore(distKey1)
 			store.Set(sKey, sVal)
 
 			expectedChangeSet = append(expectedChangeSet, &storetypes.StoreKVPair{
@@ -127,7 +127,7 @@ func Test_Ctx_with_StreamingManager(t *testing.T) {
 	})
 	require.NoError(t, err)
 
-	ctx := getFinalizeBlockStateCtx(suite.baseApp)
+	ctx := baseapptestutil.GetFinalizeBlockStateCtx(suite.baseApp)
 	sm := ctx.StreamingManager()
 	require.NotNil(t, sm, fmt.Sprintf("nil StreamingManager: %v", sm))
 	require.Equal(t, listeners, sm.ABCIListeners, fmt.Sprintf("should contain same listeners: %v", listeners))
@@ -139,7 +139,7 @@ func Test_Ctx_with_StreamingManager(t *testing.T) {
 
 		_, err = suite.baseApp.FinalizeBlock(&abci.RequestFinalizeBlock{Height: int64(blockN) + 1})
 		require.NoError(t, err)
-		ctx := getFinalizeBlockStateCtx(suite.baseApp)
+		ctx := baseapptestutil.GetFinalizeBlockStateCtx(suite.baseApp)
 		sm := ctx.StreamingManager()
 		require.NotNil(t, sm, fmt.Sprintf("nil StreamingManager: %v", sm))
 		require.Equal(t, listeners, sm.ABCIListeners, fmt.Sprintf("should contain same listeners: %v", listeners))
